Add exceptions for ListSchedule
Included EmptyListException when there is no class, test or cca. It
will prompt user to add some instead.

Added logging when listing so that we can see the control flow of
the program.

Added assertions in printArrayPaddedNumbers methods to ensure
arguments passed in is not nullable and event array size is not 0.